more cleanup and fix tiled
